+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Items with *BC!* note about backwards compatibility breaks!     
 
 ## [Unreleased]
+
+## [0.8.0] - 2017-10-31
 ## Renamed
 - *BC!*: HandlerFactory -> DefaultHandlerFactory. 
 - *BC!*: FreeElephants\RestDaemon\Endpoint\AbstractEndpointMethodHandler -> FreeElephants\RestDaemon\Endpoint\Handler\AbstractEndpointMethodHandler
@@ -107,7 +109,8 @@ Items with *BC!* note about backwards compatibility breaks!
 - Adopt to use with one of two http-daemon drivers: Ratchet (ReactPHP) and Aerys (amphp). 
 - All features. 
 
-[Unreleased]: https://github.com/FreeElephants/rest-daemon/compare/0.7.1...HEAD
+[Unreleased]: https://github.com/FreeElephants/rest-daemon/compare/0.8.0...HEAD
+[0.8.0]: https://github.com/FreeElephants/rest-daemon/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/FreeElephants/rest-daemon/compare/0.7.0...0.7.1
 [0.7.0]: https://github.com/FreeElephants/rest-daemon/compare/0.6.1...0.7.0
 [0.6.1]: https://github.com/FreeElephants/rest-daemon/compare/0.6.0...0.6.1